[build][pytools] site_uploader
